Changed SubType percentages to doubles. Changing the rounding method (in rounds on two decimals now)

+++ src/test/java/nl/halewijn/persoonlijkheidstest/domain/QuestionnaireTest.java
@@ -226,7 +226,7 @@ public class QuestionnaireTest {
 		localQuestionService.save(theoremBattle);
 		questionnaire.setAnsweredQuestions(questions);
 		double[] pTypeResultArray = questionnaire.calculatePersonalityTypeResults(questions);
-		int[] subTypeResultArray = questionnaire.calculateSubTypeResults(questions);
+		double[] subTypeResultArray = questionnaire.calculateSubTypeResults(questions);
 		assertEquals(1.0, pTypeResultArray[0], 0);
 		assertEquals(33.0, subTypeResultArray[0], 0);
